# improve performance of parse