Improves performance by preventing everything getting lifted to the Char domain.

Benchmark

```jl
function parseintperf(t)
   local n, m
   for i=1:t
       n = rand(UInt32)
       s = string(n, base = 16)
       m = UInt32(parse(Int64,s, base = 16))
       @assert m == n
   end
   return n
end


strings = [string(rand(UInt32)) for i in 1:10^4];
f(strings) = for s in strings; parse(Int, s) end;
```

Before

```jl
julia> @btime parseintperf(1000);
  189.958 μs (2999 allocations: 109.36 KiB)

julia> @btime f(strings)
  848.226 μs (0 allocations: 0 bytes)
```

After

```jl
julia> @btime parseintperf(1000)
  170.692 μs (2999 allocations: 109.36 KiB)

julia> @btime f(strings)
  709.092 μs (0 allocations: 0 bytes)
```

Still not completely caught up with 0.6 for `parseintperf` but getting close.