@MasonProtter 2018-08-20
What were the numbers for 0.6, for comparison?