@StefanKarpinski 2018-08-20
Ideally this should really be necessary since the comparison operation between characters is just an integer comparison so it seems like something else is not quite right here if this is necessary for good performance.