@KristofferC 2018-08-21
It is not the comparison that is expensive, it is the arithmetic (`-` in this case), see https://github.com/JuliaLang/julia/issues/28671.

However, hoisting the conversion like this still has a ~4% performance improvement so might as well leave it?