@KristofferC 2018-08-21
@MasonProtter 

0.6:

```
julia> @btime parseintperf(1000);
  143.515 μs (2000 allocations: 93.75 KiB)

julia> @btime f(strings)
  793.624 μs (0 allocations: 0 bytes)
```