@quinnj 2018-08-21
Parsers.jl hereby challenges you to a duel:
```julia
julia> strings = [string(rand(UInt32)) for i in 1:10^4];

julia> f(strings) = for s in strings; Parsers.parse(s, Int) end;

julia> @btime f(strings)
  549.167 μs (30000 allocations: 1.37 MiB)
```