@KristofferC 2018-08-21
```
# Base
julia> @btime f(strings)
  whatevs (0 allocations: 0 bytes)

# Parsers.jl
julia> @btime f(strings)
  whatevs (30000 allocations: 1.37 MiB)
```

Well?