@KristofferC 2018-08-22
Even though this comes with a small readability regression, I think it is worth it for such a basic function (we also show `parseintperf` on the benchmark plot on the homepage).