@KristofferC 2018-11-13
@quinnj Getting closer ;)

```jl
julia> @btime f(strings)
  595.071 μs (0 allocations: 0 bytes)
```